+++ Projects/ProblemGirkmann/tools.h
@@ -132,7 +132,7 @@ public:
 	/**
 	 * Find the indices of the corner connects
 	 */
-	void CornerConnects(TPZCompMesh *cmesh, std::set<int> &indices, int matid);
+	void CornerConnects(TPZCompMesh *cmesh, std::set<long> &indices, int matid);
 		
 protected:
